docs(screenshot): update screenshot for session 4
Previously...

I forgot to add the Screenshot of the update of the webpage for session 4.
And the Screenshot documentation wouldn't be shown in the MarkDown.

________

Now...

I have added a screenshot for session 4, and I put it inside "images/documentations/".
Also, I have fix the link for the screenshots inside the Markdown for this branch.